requirements.txt
uvicorn[standard]==0.23.2
fastapi==0.104.1
black==23.10.1
pytest==7.4.3
pytest-mock
pydantic==2.4.2


# API Client
httpx==0.25.0

# SQl Alch
SQLAlchemy==2.0.22
sqlalchemy-utils==0.41.1
psycopg2-binary==2.9.9

python-dotenv==1.0.0


pre-commit==3.5.0